conda_requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=conda_forge
_openmp_mutex=4.5=2_kmp_llvm
alsa-lib=1.2.8=h166bdaf_0
aom=3.3.0=h27087fc_1
astunparse=1.6.3=pypi_0
attr=2.5.1=h166bdaf_1
blas=1.0=mkl
bottleneck=1.3.5=py38h7deecbd_0
brotli=1.0.9=h166bdaf_9
brotli-bin=1.0.9=h166bdaf_9
brotlipy=0.7.0=py38h27cfd23_1003
bzip2=1.0.8=h7b6447c_0
c-ares=1.19.1=hd590300_0
ca-certificates=2023.7.22=hbcca054_0
cairo=1.16.0=ha61ee94_1014
certifi=2023.7.22=pyhd8ed1ab_0
cffi=1.15.1=py38h5eee18b_3
charset-normalizer=2.0.4=pyhd3eb1b0_0
colorama=0.4.6=pyhd8ed1ab_0
contourpy=1.1.0=py38h7f3f72f_0
cryptography=39.0.1=py38h9ce1e76_2
cuda-cudart=11.7.99=0
cuda-cupti=11.7.101=0
cuda-libraries=11.7.1=0
cuda-nvrtc=11.7.99=0
cuda-nvtx=11.7.91=0
cuda-runtime=11.7.1=0
cycler=0.11.0=pyhd8ed1ab_0
dbus=1.13.6=h5008d03_3
decord=0.6.0=pypi_0
dm-tree=0.1.8=pypi_0
efficientnet-pytorch=0.7.1=pyhd8ed1ab_1
einops=0.6.1=pyhd8ed1ab_0
expat=2.5.0=hcb278e6_1
ffmpeg=4.4.1=h964e5f1_4
fftw=3.3.10=nompi_hc118613_108
filelock=3.9.0=py38h06a4308_0
font-ttf-dejavu-sans-mono=2.37=hab24e00_0
font-ttf-inconsolata=3.000=h77eed37_0
font-ttf-source-code-pro=2.038=h77eed37_0
font-ttf-ubuntu=0.83=hab24e00_0
fontconfig=2.14.2=h14ed4e7_0
fonts-conda-ecosystem=1=0
fonts-conda-forge=1=0
fonttools=4.41.0=py38h01eb140_0
freeglut=3.2.2=h9c3ff4c_1
freetype=2.12.1=h4a9f257_0
fsspec=2023.6.0=pyh1a96a4e_0
gast=0.5.4=pypi_0
gettext=0.21.1=h27087fc_0
giflib=5.2.1=h5eee18b_3
glib=2.76.4=hfc55251_0
glib-tools=2.76.4=hfc55251_0
gmp=6.2.1=h295c915_3
gmpy2=2.1.2=py38heeb90bb_0
gnutls=3.6.15=he1e5248_0
graphite2=1.3.13=h58526e2_1001
gst-plugins-base=1.21.3=h4243ec0_1
gstreamer=1.21.3=h25f0c4b_1
gstreamer-orc=0.4.34=hd590300_0
harfbuzz=4.4.1=hf9f4e7c_0
hdf5=1.12.1=nompi_h4df4325_104
huggingface_hub=0.16.4=pyhd8ed1ab_0
icu=70.1=h27087fc_0
idna=3.4=py38h06a4308_0
imageio=2.26.0=pyh24c5eb1_0
importlib-metadata=6.8.0=pyha770c72_0
importlib-resources=6.0.0=pyhd8ed1ab_1
importlib_resources=6.0.0=pyhd8ed1ab_1
intel-openmp=2023.1.0=hdb19cb5_46305
jack=1.9.22=h11f4161_0
jasper=2.0.33=h0ff4b12_1
jinja2=3.1.2=py38h06a4308_0
joblib=1.1.1=py38h06a4308_0
jpeg=9e=h5eee18b_1
jsonpatch=1.32=pyhd8ed1ab_0
jsonpointer=2.0=py_0
keyutils=1.6.1=h166bdaf_0
kiwisolver=1.4.4=py38h43d8883_1
krb5=1.20.1=h81ceb04_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libblas=3.9.0=1_h86c2bf4_netlib
libbrotlicommon=1.0.9=h166bdaf_9
libbrotlidec=1.0.9=h166bdaf_9
libbrotlienc=1.0.9=h166bdaf_9
libcap=2.67=he9d0100_0
libcblas=3.9.0=5_h92ddd45_netlib
libclang=15.0.7=default_h7634d5b_2
libclang13=15.0.7=default_h9986a30_2
libcublas=11.10.3.66=0
libcufft=10.7.2.124=h4fbf590_0
libcufile=1.7.0.149=0
libcups=2.3.3=h36d4200_3
libcurand=10.3.3.53=0
libcurl=8.1.2=h409715c_0
libcusolver=11.4.0.1=0
libcusparse=11.7.4.91=0
libdb=6.2.32=h9c3ff4c_0
libdeflate=1.8=h7f98852_0
libdrm=2.4.114=h166bdaf_0
libedit=3.1.20191231=he28a2e2_2
libev=4.33=h516909a_1
libevent=2.1.10=h28343ad_4
libexpat=2.5.0=hcb278e6_1
libffi=3.4.4=h6a678d5_0
libflac=1.4.3=h59595ed_0
libgcc-ng=13.1.0=he5830b7_0
libgcrypt=1.10.1=h166bdaf_0
libgfortran-ng=13.1.0=h69a702a_0
libgfortran5=13.1.0=h15d22d2_0
libglib=2.76.4=hebfc3b9_0
libglu=9.0.0=he1b5a44_1001
libgpg-error=1.47=h71f35ed_0
libiconv=1.17=h166bdaf_0
libidn2=2.3.4=h5eee18b_0
liblapack=3.9.0=5_h92ddd45_netlib
liblapacke=3.9.0=5_h92ddd45_netlib
libllvm15=15.0.7=hadd5161_1
libnghttp2=1.52.0=h61bc06f_0
libnpp=11.7.4.75=0
libnvjpeg=11.8.0.2=0
libogg=1.3.4=h7f98852_1
libopencv=4.5.5=py38hc65905f_11
libopus=1.3.1=h7f98852_1
libpciaccess=0.17=h166bdaf_0
libpng=1.6.39=h5eee18b_0
libpq=15.3=hbcd7760_1
libprotobuf=3.20.3=h3eb15da_0
libsndfile=1.2.0=hb75c966_0
libsqlite=3.42.0=h2797004_0
libssh2=1.11.0=h0841786_0
libstdcxx-ng=13.1.0=hfd8a6a1_0
libsystemd0=253=h8c4010b_1
libtasn1=4.19.0=h5eee18b_0
libtiff=4.4.0=hecacb30_2
libtool=2.4.7=h27087fc_0
libudev1=253=h0b41bf4_1
libunistring=0.9.10=h27cfd23_0
libuuid=2.38.1=h0b41bf4_0
libva=2.18.0=h0b41bf4_0
libvorbis=1.3.7=h9c3ff4c_0
libvpx=1.11.0=h9c3ff4c_3
libwebp=1.2.4=h11a3e52_1
libwebp-base=1.2.4=h5eee18b_1
libxcb=1.13=h7f98852_1004
libxkbcommon=1.5.0=h79f4944_0
libxml2=2.10.3=hca2bb57_4
libzlib=1.2.13=hd590300_5
llvm-openmp=14.0.6=h9e868ea_0
lz4-c=1.9.4=h6a678d5_0
markupsafe=2.1.1=py38h7f8727e_0
matplotlib=3.7.2=py38h578d9bd_0
matplotlib-base=3.7.2=py38hf5b0b65_0
mkl=2023.1.0=h6d00ec8_46342
mkl-service=2.4.0=py38h5eee18b_1
mkl_fft=1.3.6=py38h417a72b_1
mkl_random=1.2.2=py38h417a72b_1
mpc=1.1.0=h10f8cd9_1
mpfr=4.0.2=hb69a4c5_1
mpg123=1.31.3=hcb278e6_0
mpmath=1.2.1=py38h06a4308_0
munch=4.0.0=pyhd8ed1ab_0
munkres=1.1.4=pyh9f0ad1d_0
mysql-common=8.0.33=hf1915f5_2
mysql-libs=8.0.33=hca2cd23_2
ncurses=6.4=h6a678d5_0
nettle=3.7.3=hbbd107a_1
networkx=2.8.4=py38h06a4308_1
nspr=4.35=h27087fc_0
nss=3.89=he45b914_0
numexpr=2.8.4=py38hc78ab66_1
numpy=1.24.3=py38hf6e8229_1
numpy-base=1.24.3=py38h060ed82_1
nvidia-dali-cuda110=1.27.0=pypi_0
opencv=4.5.5=py38h578d9bd_11
openh264=2.1.1=h4ff587b_0
openssl=3.1.1=hd590300_1
packaging=23.1=pyhd8ed1ab_0
pandas=1.5.3=py38h417a72b_0
pcre2=10.40=hc3806b6_0
pillow=9.4.0=py38h6a678d5_0
pip=23.1.2=py38h06a4308_0
pixman=0.40.0=h36c2ea0_0
platformdirs=3.9.1=pyhd8ed1ab_0
ply=3.11=py_1
pooch=1.7.0=pyha770c72_3
pretrainedmodels=0.7.4=pyhd8ed1ab_2
protobuf=3.20.3=py38h8dc9893_1
pthread-stubs=0.4=h36c2ea0_1001
pulseaudio=16.1=hcb278e6_3
pulseaudio-client=16.1=h5195f5e_3
pulseaudio-daemon=16.1=ha8d29e2_3
py-opencv=4.5.5=py38h7f3c49e_11
pycparser=2.21=pyhd3eb1b0_0
pyopenssl=23.0.0=py38h06a4308_0
pyparsing=3.0.9=pyhd8ed1ab_0
pyqt=5.15.7=py38ha0d8c90_3
pyqt5-sip=12.11.0=py38h8dc9893_3
pysocks=1.7.1=py38h06a4308_0
python=3.8.17=h955ad1f_0
python-dateutil=2.8.2=pyhd8ed1ab_0
python_abi=3.8=2_cp38
pytorch=2.0.1=py3.8_cuda11.7_cudnn8.5.0_0
pytorch-cuda=11.7=h778d358_5
pytorch-mutex=1.0=cuda
pytz=2022.7=py38h06a4308_0
pyyaml=6.0=py38h0a891b7_5
qt-main=5.15.6=hf6cd601_5
readline=8.2=h5eee18b_0
requests=2.29.0=py38h06a4308_0
safetensors=0.3.1=py38h0cc4f7c_0
scikit-learn=1.2.0=py38h6a678d5_0
scipy=1.10.1=py38hf6e8229_1
segmentation-models-pytorch=0.3.3=pyhd8ed1ab_0
setuptools=67.8.0=py38h06a4308_0
sip=6.7.9=py38h17151c0_0
six=1.16.0=pyh6c4a22f_0
sqlite=3.41.2=h5eee18b_0
svt-av1=1.1.0=h27087fc_1
sympy=1.11.1=py38h06a4308_0
tbb=2021.8.0=hdb19cb5_0
tensorboardx=2.2=pyhd3eb1b0_0
threadpoolctl=2.2.0=pyh0d69192_0
timm=0.9.2=pyhd8ed1ab_0
tk=8.6.12=h1ccaba5_0
toml=0.10.2=pyhd8ed1ab_0
tomli=2.0.1=pyhd8ed1ab_0
torchaudio=2.0.2=py38_cu117
torchinfo=1.8.0=pyhd8ed1ab_0
torchtriton=2.0.0=py38
torchvision=0.15.2=py38_cu117
tornado=6.3.2=py38h01eb140_0
tqdm=4.65.0=pyhd8ed1ab_1
typing-extensions=4.6.3=py38h06a4308_0
typing_extensions=4.6.3=py38h06a4308_0
unicodedata2=15.0.0=py38h0a891b7_0
urllib3=1.26.16=py38h06a4308_0
visdom=0.2.4=pyhd8ed1ab_0
websocket-client=1.6.1=pyhd8ed1ab_0
wheel=0.38.4=py38h06a4308_0
x264=1!161.3030=h7f98852_1
x265=3.5=h924138e_3
xcb-util=0.4.0=h166bdaf_0
xcb-util-image=0.4.0=h166bdaf_0
xcb-util-keysyms=0.4.0=h166bdaf_0
xcb-util-renderutil=0.3.9=h166bdaf_0
xcb-util-wm=0.4.1=h166bdaf_0
xorg-fixesproto=5.0=h7f98852_1002
xorg-inputproto=2.3.2=h7f98852_1002
xorg-kbproto=1.0.7=h7f98852_1002
xorg-libice=1.1.1=hd590300_0
xorg-libsm=1.2.4=h7391055_0
xorg-libx11=1.8.4=h0b41bf4_0
xorg-libxau=1.0.11=hd590300_0
xorg-libxdmcp=1.1.3=h7f98852_0
xorg-libxext=1.3.4=h0b41bf4_2
xorg-libxfixes=5.0.3=h7f98852_1004
xorg-libxi=1.7.10=h7f98852_0
xorg-libxrender=0.9.10=h7f98852_1003
xorg-renderproto=0.11.1=h7f98852_1002
xorg-xextproto=7.3.0=h0b41bf4_1003
xorg-xproto=7.0.31=h7f98852_1007
xz=5.4.2=h5eee18b_0
yaml=0.2.5=h7f98852_2
zipp=3.16.2=pyhd8ed1ab_0
zlib=1.2.13=hd590300_5
zstd=1.5.5=hc292b87_0
